Rebuild the number functions to store the number internally as a string.
This makes the output of the encoder more predictable and the value will only
be decoded into an integer as it is used. The API for encoding a number from
a double is still in place, but two additional functions have been added;
json_create_number_str() which just stores the string (but with the JSON_NUMBER
type) and json_number() which takes a JSON_NUMBER item and decodes the string
into a double.